// src/lib/pb/db-types.ts
export type UtilityStaffResponse = {
  name: string;
  type: 'caretaker' | 'manager' | 'cashier';
  avatar?: string;
} & AuthCollectionRecord;

export type UtilityStaffCreate = {
  name: string;
  type: 'caretaker' | 'manager' | 'cashier';
  avatar?: string;
};

export type UtilityStaffUpdate = {
  name?: string;
  type?: 'caretaker' | 'manager' | 'cashier';
  avatar?: string;
};

export type UtilityStaffCollection = {
  type: 'auth';
  collectionId: 'dlvnttlfiw585jv';
  collectionName: 'utility_staff';
  response: UtilityStaffResponse;
  create: UtilityStaffCreate;
  update: UtilityStaffUpdate;
  relations: {
    'tasky_tasks(created_by)': TaskyTasksCollection[];
    'tasky_tasks(updated_by)': TaskyTasksCollection[];
    'tasky_tasks(approved_by)': TaskyTasksCollection[];
    'tasky_tasks(funded_by)': TaskyTasksCollection[];
    'tasky_tasks(marked_completed_by)': TaskyTasksCollection[];
    'tasky_tasks(rejected_by)': TaskyTasksCollection[];
    'tasky_tasks(marked_in_progress_by)': TaskyTasksCollection[];
    'tasky_staff_details(leave_requested_by)': TaskyStaffDetailsCollection[];
    'tasky_staff_details(leave_approved_by)': TaskyStaffDetailsCollection[];
    'pocketbook_notifications(source)': PocketbookNotificationsCollection[];
  };
};
// ...
export type TaskyTasksCollection = {
  type: 'base';
  collectionId: 'ipamtot35z4hzln';
  collectionName: 'tasky_tasks';
  response: TaskyTasksResponse;
  create: TaskyTasksCreate;
  update: TaskyTasksUpdate;
  relations: {
    created_by: UtilityStaffCollection;
    updated_by: UtilityStaffCollection;
    approved_by: UtilityStaffCollection;
    funded_by: UtilityStaffCollection;
    marked_completed_by: UtilityStaffCollection;
    rejected_by: UtilityStaffCollection;
    marked_in_progress_by: UtilityStaffCollection;
  };
};
// ...
export type TaskyStaffDetailsCollection = {
  type: 'base';
  collectionId: 'm1ibf55enmz09s6';
  collectionName: 'tasky_staff_details';
  response: TaskyStaffDetailsResponse;
  create: TaskyStaffDetailsCreate;
  update: TaskyStaffDetailsUpdate;
  relations: {
    leave_requested_by: UtilityStaffCollection;
    leave_approved_by: UtilityStaffCollection;
  };
};
// ...
export type PocketbookNotificationsCollection = {
  type: 'base';
  collectionId: 'q2inux1nn05ynut';
  collectionName: 'pocketbook_notifications';
  response: PocketbookNotificationsResponse;
  create: PocketbookNotificationsCreate;
  update: PocketbookNotificationsUpdate;
  relations: {
    source: UtilityStaffCollection;
  };
};

// ===== utility_shops =====

export type UtilityShopsResponse = {
  shop_number: string;
  supa_tenant?: string;
  utils?: 'elec' | 'water' | 'both' | 'none';
  order?: number;
  supa_id?: string;
  is_vacant?: boolean;
  tenant?: string;
} & BaseCollectionRecord;

export type UtilityShopsCreate = {
  shop_number: string;
  supa_tenant?: string;
  utils?: 'elec' | 'water' | 'both' | 'none';
  order?: number;
  supa_id?: string;
  is_vacant?: boolean;
  tenant?: string;
};

export type UtilityShopsUpdate = {
  shop_number?: string;
  supa_tenant?: string;
  utils?: 'elec' | 'water' | 'both'|'none';
  order?: number;
  'order+'?: number;
  'order-'?: number;
  supa_id?: string;
  is_vacant?: boolean;
  tenant?: string;
};

export type UtilityShopsCollection = {
  type: 'base';
  collectionId: 'zb1etrv0i3olw5p';
  collectionName: 'utility_shops';
  response: UtilityShopsResponse;
  create: UtilityShopsCreate;
  update: UtilityShopsUpdate;
  relations: {
    tenant: UtilityTenantsCollection;
    'utility_bills(shop)': UtilityBillsCollection[];
  };
};

// ===== utility_tenants_base =====

export type UtilityTenantsBaseResponse = {
  name: string;
  contact?: string;
  email?: string;
  details?: string;
  supa_id?: string;
} & BaseCollectionRecord;

export type UtilityTenantsBaseCreate = {
  name: string;
  contact?: string;
  email?: string;
  details?: string;
  supa_id?: string;
};

export type UtilityTenantsBaseUpdate = {
  name?: string;
  contact?: string;
  email?: string;
  details?: string;
  supa_id?: string;
};

export type UtilityTenantsBaseCollection = {
  type: 'base';
  collectionId: '6ur1ivky21zygnv';
  collectionName: 'utility_tenants_base';
  response: UtilityTenantsBaseResponse;
  create: UtilityTenantsBaseCreate;
  update: UtilityTenantsBaseUpdate;
  relations: {};
};

// ===== utility_bills =====

export type UtilityBillsResponse = {
  shop: string;
  elec_readings?: number;
  water_readings?: number;
  month: number;
  year: number;
} & BaseCollectionRecord;

export type UtilityBillsCreate = {
  shop: string;
  elec_readings?: number;
  water_readings?: number;
  month: number;
  year: number;
};

export type UtilityBillsUpdate = {
  shop?: string;
  elec_readings?: number;
  'elec_readings+'?: number;
  'elec_readings-'?: number;
  water_readings?: number;
  'water_readings+'?: number;
  'water_readings-'?: number;
  month?: number;
  'month+'?: number;
  'month-'?: number;
  year?: number;
  'year+'?: number;
  'year-'?: number;
};

export type UtilityBillsCollection = {
  type: 'base';
  collectionId: 'cvtakohtxaagiat';
  collectionName: 'utility_bills';
  response: UtilityBillsResponse;
  create: UtilityBillsCreate;
  update: UtilityBillsUpdate;
  relations: {
    shop: UtilityShopsCollection;
  };
};
// ...
export type UtilityTenantsResponse = {
  phone?: string;
  avatar?:string; 
} & AuthCollectionRecord;

export type UtilityTenantsCreate = {
  phone?: string;
  avatar?: string|File;
};

export type UtilityTenantsUpdate = {
  phone?: string;
  avatar?: string | File;
};

export type UtilityTenantsCollection = {
  type: 'auth';
  collectionId: 'vdjzfr0gybtmwif';
  collectionName: 'utility_tenants';
  response: UtilityTenantsResponse;
  create: UtilityTenantsCreate;
  update: UtilityTenantsUpdate;
  relations: {
    'utility_shops(tenant)': UtilityShopsCollection[];
  };
};

// ===== Schema =====

export type Schema = {
// mashamba
  // mashamba_listings: MashambaListingsCollection;
  // mashamba_owner: MashambaOwnerCollection;
// tasky
  // tasky_tasks: TaskyTasksCollection;
  // tasky_staff_details: TaskyStaffDetailsCollection;
// pocketbook
  // pocketbook_user: PocketbookUserCollection;
  // pocketbook_reactions: PocketbookReactionsCollection;
  // pocketbook_notifications: PocketbookNotificationsCollection;
  // pocketbook_posts: PocketbookPostsCollection;
  // pocketbook_friends: PocketbookFriendsCollection;
// utility
  utility_shops: UtilityShopsCollection;
  utility_staff: UtilityStaffCollection;
  utility_tenants_base: UtilityTenantsBaseCollection;
  utility_bills: UtilityBillsCollection;
  utility_tenants: UtilityTenantsCollection;
};